+++ hydra-node/test/Hydra/BehaviorSpec.hs
@@ -22,7 +22,7 @@ import Data.List qualified as List
 import Hydra.API.ClientInput
 import Hydra.API.Server (Server (..), mkTimedServerOutputFromStateEvent)
 import Hydra.API.ServerOutput (ClientMessage (..), DecommitInvalidReason (..), ServerOutput (..), TimedServerOutput (..))
-import Hydra.Cardano.Api (ChainPoint (..), SigningKey)
+import Hydra.Cardano.Api (SigningKey)
 import Hydra.Chain (
   Chain (..),
   ChainEvent (..),
@@ -30,13 +30,13 @@ import Hydra.Chain (
   PostChainTx (..),
   initHistory,
  )
-import Hydra.Chain.ChainState (ChainStateType, IsChainState, chainStatePoint)
+import Hydra.Chain.ChainState (ChainSlot (ChainSlot), ChainStateType, IsChainState, chainStateSlot)
 import Hydra.Chain.Direct.Handlers (LocalChainState, getLatest, newLocalChainState, pushNew, rollback)
 import Hydra.Events (EventSink (..))
 import Hydra.Events.Rotation (EventStore (..))
 import Hydra.HeadLogic (CoordinatedHeadState (..), Effect (..), HeadState (..), InitialState (..), Input (..), OpenState (..))
 import Hydra.HeadLogicSpec (testSnapshot)
-import Hydra.Ledger (Ledger, nextChainPoint)
+import Hydra.Ledger (Ledger, nextChainSlot)
 import Hydra.Ledger.Simple (SimpleChainState (..), SimpleTx (..), aValidTx, simpleLedger, utxoRef, utxoRefs)
 import Hydra.Logging (Tracer)
 import Hydra.Network (Network (..))
@@ -69,7 +69,6 @@ import Hydra.Tx.Crypto (HydraKey, aggregate, sign)
 import Hydra.Tx.IsTx (IsTx (..))
 import Hydra.Tx.Party (Party (..), deriveParty, getParty)
 import Hydra.Tx.Snapshot (Snapshot (..), SnapshotNumber, getSnapshot)
-import System.Random (split)
 import Test.Hydra.Tx.Fixture (
   alice,
   aliceSk,
@@ -80,8 +79,6 @@ import Test.Hydra.Tx.Fixture (
   testHeadSeed,
  )
 import Test.QuickCheck (chooseEnum, counterexample, forAll, getNegative, ioProperty)
-import Test.QuickCheck.Gen (unGen)
-import Test.QuickCheck.Random (QCGen, mkQCGen)
 import Test.Util (
   propRunInSim,
   shouldBe,
@@ -892,8 +889,8 @@ spec = parallel $ do
               -- XXX: This is a bit cumbersome and maybe even incorrect (chain
               -- states), the simulated chain should provide a way to inject an
               -- 'OnChainTx' without providing a chain state?
-              injectChainEvent n1 Observation{observedTx = OnCloseTx testHeadId 0 deadline, newChainState = SimpleChainState{point = ChainPointAtGenesis}}
-              injectChainEvent n2 Observation{observedTx = OnCloseTx testHeadId 0 deadline, newChainState = SimpleChainState{point = ChainPointAtGenesis}}
+              injectChainEvent n1 Observation{observedTx = OnCloseTx testHeadId 0 deadline, newChainState = SimpleChainState{slot = ChainSlot 0}}
+              injectChainEvent n2 Observation{observedTx = OnCloseTx testHeadId 0 deadline, newChainState = SimpleChainState{slot = ChainSlot 0}}
 
               waitUntilMatch [n1, n2] $ \case
                 HeadIsClosed{snapshotNumber} -> guard $ snapshotNumber == 0
@@ -1068,7 +1065,7 @@ withSimulatedChainAndNetwork ::
   m a
 withSimulatedChainAndNetwork =
   bracket
-    (simulatedChainAndNetwork SimpleChainState{point = ChainPointAtGenesis})
+    (simulatedChainAndNetwork SimpleChainState{slot = ChainSlot 0})
     (cancel . tickThread)
 
 -- | Creates a simulated chain and network to which 'HydraNode's can be
@@ -1086,7 +1083,6 @@ simulatedChainAndNetwork initialChainState = do
   nextTxId <- newLabelledTVarIO "sim-chain-next-txid" 10000
   localChainState <- newLocalChainState (initHistory initialChainState)
   tickThread <- asyncLabelled "sim-chain-tick" $ simulateTicks nodes localChainState
-  rngVar <- newLabelledTVarIO "run-gen-stm" $ mkQCGen 42
   pure $
     SimulatedChainNetwork
       { connectNode = \draftNode -> do
@@ -1098,7 +1094,7 @@ simulatedChainAndNetwork initialChainState = do
                       -- Only observe "after one block"
                       void . asyncLabelled "sim-chain-post-tx" $ do
                         threadDelay blockTime
-                        createAndYieldEvent rngVar nodes history localChainState $ toOnChainTx now tx
+                        createAndYieldEvent nodes history localChainState $ toOnChainTx now tx
                   , draftCommitTx = \_ -> error "unexpected call to draftCommitTx"
                   , draftDepositTx = \_ -> error "unexpected call to draftDepositTx"
                   , submitTx = \_ -> error "unexpected call to submitTx"
@@ -1113,11 +1109,11 @@ simulatedChainAndNetwork initialChainState = do
       , tickThread
       , rollbackAndForward = rollbackAndForward nodes history localChainState
       , simulateCommit = \headId party toCommit ->
-          createAndYieldEvent rngVar nodes history localChainState $ OnCommitTx{headId, party, committed = toCommit}
+          createAndYieldEvent nodes history localChainState $ OnCommitTx{headId, party, committed = toCommit}
       , simulateDeposit = \headId toDeposit deadline -> do
           created <- getCurrentTime
           depositTxId <- atomically $ stateTVar nextTxId (\i -> (i, i + 1))
-          createAndYieldEvent rngVar nodes history localChainState $
+          createAndYieldEvent nodes history localChainState $
             OnDepositTx{headId, deposited = toDeposit, created, deadline, depositTxId}
           pure depositTxId
       , closeWithInitialSnapshot = error "unexpected call to closeWithInitialSnapshot"
@@ -1131,15 +1127,14 @@ simulatedChainAndNetwork initialChainState = do
     now <- getCurrentTime
     event <- atomically $ do
       cs <- getLatest localChainState
-      let chainPoint = chainStatePoint cs
-      let lastKnown = SimpleChainState chainPoint
-      pure $ Tick now lastKnown
+      -- XXX: This chain state (its slot) does not correspond to 'now'
+      pure $ Tick now cs
     readTVarIO nodes >>= mapM_ (`handleChainEvent` event)
 
-  createAndYieldEvent rngVar nodes history localChainState tx = do
+  createAndYieldEvent nodes history localChainState tx = do
     chainEvent <- atomically $ do
       cs <- getLatest localChainState
-      cs' <- advancePoint rngVar cs
+      let cs' = advanceSlot cs
       pushNew localChainState cs'
       pure $
         Observation
@@ -1148,13 +1143,7 @@ simulatedChainAndNetwork initialChainState = do
           }
     recordAndYieldEvent nodes history chainEvent
 
-  advancePoint ::
-    TVar m QCGen ->
-    SimpleChainState ->
-    STM m SimpleChainState
-  advancePoint rngVar SimpleChainState{point} = do
-    nextPoint <- runGenSTM rngVar (nextChainPoint point)
-    pure SimpleChainState{point = nextPoint}
+  advanceSlot SimpleChainState{slot} = SimpleChainState{slot = nextChainSlot slot}
 
   recordAndYieldEvent ::
     TVar m [HydraNode tx m] ->
@@ -1169,6 +1158,7 @@ simulatedChainAndNetwork initialChainState = do
       handleChainEvent n chainEvent
 
   rollbackAndForward ::
+    IsChainState tx =>
     TVar m [HydraNode tx m] ->
     TVar m [ChainEvent tx] ->
     LocalChainState m tx ->
@@ -1181,15 +1171,15 @@ simulatedChainAndNetwork initialChainState = do
       writeTVar history kept
       pure (reverse toReplay, kept)
     -- Determine the new (last kept one) chainstate
-    let chainState =
+    let chainSlot =
           List.head $
             map
               ( \case
-                  Observation{newChainState} -> newChainState
+                  Observation{newChainState} -> chainStateSlot newChainState
                   _NoObservation -> error "unexpected non-observation ChainEvent"
               )
               kept
-    rolledBackChainState <- atomically $ rollback localChainState chainState
+    rolledBackChainState <- atomically $ rollback localChainState chainSlot
     -- Yield rollback events
     ns <- readTVarIO nodes
     chainTime <- getCurrentTime
@@ -1197,15 +1187,6 @@ simulatedChainAndNetwork initialChainState = do
     -- Re-play the observation events
     forM_ toReplay $ \ev ->
       recordAndYieldEvent nodes history ev
-
-runGenSTM :: MonadSTM m => TVar m QCGen -> Gen a -> STM m a
-runGenSTM genVar gen = do
-  g <- readTVar genVar
-  let (g1, g2) = split g
-  -- 30 is a common default QC "size"
-  let x = unGen gen g1 30
-  writeTVar genVar g2
-  pure x
 
 handleChainEvent :: HydraNode tx m -> ChainEvent tx -> m ()
 handleChainEvent HydraNode{inputQueue} = enqueue inputQueue . ChainInput
@@ -1300,7 +1281,7 @@ withHydraNode' dp signingKey otherParties chain action = do
   outputs <- newLabelledTQueueIO "hydra-node-outputs"
   messages <- newLabelledTQueueIO "hydra-node-messages"
   outputHistory <- newLabelledTVarIO "hydra-node-output-history" mempty
-  let initialChainState = SimpleChainState{point = ChainPointAtGenesis}
+  let initialChainState = SimpleChainState{slot = ChainSlot 0}
   node@HydraNode{nodeStateHandler = NodeStateHandler{queryNodeState}} <-
     createHydraNode
       traceInIOSim